Redesigned Rom Dialog is active

+++ OSX/MacGUI/RomDialogController.swift
@@ -15,16 +15,28 @@ class RomDialogController : UserDialogController {
     let romImageLight = NSImage.init(named: NSImage.Name.init(rawValue: "rom_light"))
     let romImageMedium = NSImage.init(named: NSImage.Name.init(rawValue: "rom_medium"))
 
-    // @IBOutlet weak var learnMore: NSTextField!
-    @IBOutlet weak var kernalRom: NSImageView!
-    @IBOutlet weak var kernalRomText: NSTextField!
+    @IBOutlet weak var headerImage: NSImageView!
+    @IBOutlet weak var headerText: NSTextField!
+    @IBOutlet weak var headerSubText: NSTextField!
+
     @IBOutlet weak var basicRom: NSImageView!
     @IBOutlet weak var basicRomText: NSTextField!
+    @IBOutlet weak var basicRomButton: NSButton!
+
+    @IBOutlet weak var kernalRom: NSImageView!
+    @IBOutlet weak var kernalRomText: NSTextField!
+    @IBOutlet weak var kernelRomButton: NSButton!
+    
     @IBOutlet weak var characterRom: NSImageView!
     @IBOutlet weak var characterRomText: NSTextField!
+    @IBOutlet weak var characterRomButton: NSButton!
+
     @IBOutlet weak var vc1541Rom: NSImageView!
     @IBOutlet weak var vc1541RomText: NSTextField!
+    @IBOutlet weak var vc1541RomButton: NSButton!
 
+    @IBOutlet weak var okButton: NSButton!
+    
     override func awakeFromNib()
     {
         refresh()
@@ -32,19 +44,69 @@ class RomDialogController : UserDialogController {
 
     func refresh()
     {
-        kernalRom.image = c64.isKernalRomLoaded() ? romImage : romImageLight
-        kernalRomText.isHidden = c64.isKernalRomLoaded()
-        basicRom.image = c64.isBasicRomLoaded() ? romImage : romImageLight
-        basicRomText.isHidden = c64.isBasicRomLoaded()
-        characterRom.image = c64.isCharRomLoaded() ? romImage : romImageLight
-        characterRomText.isHidden = c64.isCharRomLoaded()
-        vc1541Rom.image = c64.isVC1541RomLoaded() ? romImage : romImageLight
-        vc1541RomText.isHidden = c64.isVC1541RomLoaded()
+        let basic = c64.isBasicRomLoaded()
+        let kernal = c64.isKernalRomLoaded()
+        let character = c64.isCharRomLoaded()
+        let vc1541 = c64.isVC1541RomLoaded()
+
+        let readyToRun = c64.isRunnable()
+
+        headerImage.isHidden = readyToRun
+        headerText.isHidden = readyToRun
+        headerSubText.isHidden = readyToRun
+        
+        basicRom.image = basic ? romImage : romImageLight
+        basicRomText.isHidden = basic
+        basicRomButton.isHidden = !basic
+        
+        kernalRom.image = kernal ? romImage : romImageLight
+        kernalRomText.isHidden = kernal
+        kernelRomButton.isHidden = !kernal
+
+        characterRom.image = character ? romImage : romImageLight
+        characterRomText.isHidden = character
+        characterRomButton.isHidden = !character
+
+        vc1541Rom.image = vc1541 ? romImage : romImageLight
+        vc1541RomText.isHidden = vc1541
+        vc1541RomButton.isHidden = !vc1541
+        
+        okButton.stringValue = readyToRun ? "OK" : "Quit"
     }
 
+    
     //
     // Action methods
     //
+    
+    @IBAction func deleteBasicRom(_ sender: Any!)
+    {
+        track()
+        parent.c64.mem.deleteBasicRom()
+        refresh()
+    }
+    
+    @IBAction func deleteCharacterRom(_ sender: Any!)
+    {
+        track()
+        parent.c64.mem.deleteCharacterRom()
+        refresh()
+    }
+    
+    @IBAction func deleteKernalRom(_ sender: Any!)
+    {
+        track()
+        parent.c64.mem.deleteKernalRom()
+        refresh()
+    }
+    
+    @IBAction func deleteVC1541Rom(_ sender: Any!)
+    {
+        track()
+        parent.c64.drive1.deleteRom()
+        parent.c64.drive2.deleteRom()
+        refresh()
+    }
     
     @IBAction func helpAction(_ sender: Any!)
     {
@@ -53,11 +115,14 @@ class RomDialogController : UserDialogController {
         }
     }
     
-    @IBAction func quitAction(_ sender: Any!)
+    @IBAction func okAction(_ sender: Any!)
     {
         track()
         hideSheet()
-        NSApp.terminate(self)
+        
+        if (!c64.isRunnable()) {
+           NSApp.terminate(self)
+        }
     }
     
 }
